share setupCluster func
